Refactor app into separate modules
