+++ C++_VR_App.cpp
@@ -16,31 +16,26 @@
 #include <iostream>     // Standard I/O operations for console output
 #include <fstream>      // File stream handling for reading audio files
 #include <string>       // String manipulation for text processing
-#include <curl/curl.h>  // CURL operations to handle HTTP requests to APIs
+
 #include <cstdlib>      // Standard library functions, e.g., exit()
 #include <unistd.h>     // For access() function to check file existence
 #include <sstream>      // String stream operations for string manipulation
-#include <map>          // Map container for key-value pairs
+
 #include <regex>        // Regular expressions for pattern matching
-#include <vector>       // Vector container for dynamic arrays
-#include <iomanip>      // Input/output manipulators for formatting
-#include <chrono>       // For timestamp generation
+
+
+
 
 // Include the nlohmann/json library for JSON parsing and manipulation
 #include "nlohmann/json.hpp"
 // Include the configuration file for API keys
 #include "config.h"
+#include "transcription.h"
+#include "analysis.h"
+#include "notion.h"
 using json = nlohmann::json;
 using namespace std;
 
-/**
- * Global variables to store the property names from the Notion database
- * 
- * These variables are used to map the fields from the GPT response
- * to the property names in the Notion database, which may have different names.
- */
-string g_titlePropertyName = "";
-map<string, string> g_propertyNameMap;
 
 /**
  * Function to prompt the user to select an audio file
@@ -69,623 +64,6 @@ string getFileFromDialog() {
     return filePath;
 }
 
-/**
- * Callback function for CURL to write received data
- * 
- * This function is called by CURL when data is received from an HTTP request.
- * It appends the received data to the output string.
- * 
- * @param contents Pointer to the received data
- * @param size Size of each data element
- * @param nmemb Number of data elements
- * @param output Pointer to the string where data will be stored
- * @return The total size of the data received
- */
-size_t WriteCallback(void *contents, size_t size, size_t nmemb, string *output) {
-    size_t totalSize = size * nmemb;
-    output->append(reinterpret_cast<char*>(contents), totalSize);
-    return totalSize;
-}
-
-/**
- * Function to transcribe audio using the OpenAI Whisper API
- * 
- * This function sends an audio file to the OpenAI Whisper API for transcription.
- * It handles:
- * 1. Reading the audio file
- * 2. Setting up the API request with proper headers and parameters
- * 3. Sending the request to the API
- * 4. Receiving and returning the response
- * 
- * @param filePath Path to the audio file to transcribe
- * @param apiKey OpenAI API key for authentication
- * @return The API response containing the transcription
- */
-string transcribeAudio(const string &filePath, const string &apiKey) {
-    CURL *curl;
-    CURLcode res;
-    string responseString;
-    curl_global_init(CURL_GLOBAL_DEFAULT);
-    curl = curl_easy_init();
-    if (curl) {
-        ifstream file(filePath, ios::binary);
-        if (!file.is_open()) {
-            cerr << "Failed to open file: " << filePath << endl;
-            exit(EXIT_FAILURE);
-        }
-        ostringstream oss;
-        oss << file.rdbuf();
-        string fileData = oss.str();
-
-        // Set up the header with the API key
-        struct curl_slist *headers = nullptr;
-        headers = curl_slist_append(headers, ("Authorization: Bearer " + apiKey).c_str());
-
-        // Set the API endpoint for transcription
-        curl_easy_setopt(curl, CURLOPT_URL, "https://api.openai.com/v1/audio/transcriptions");
-        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-        curl_easy_setopt(curl, CURLOPT_POST, 1L);
-
-        // Set up MIME form for file upload and model parameter
-        curl_mime *form = curl_mime_init(curl);
-
-        // Add the file field to the form
-        curl_mimepart *field = curl_mime_addpart(form);
-        curl_mime_name(field, "file");
-        curl_mime_filename(field, filePath.c_str());
-        curl_mime_data(field, fileData.c_str(), fileData.size());
-
-        // Add the model parameter
-        field = curl_mime_addpart(form);
-        curl_mime_name(field, "model");
-        curl_mime_data(field, "whisper-1", CURL_ZERO_TERMINATED);
-
-        // Attach the form and set callback for response
-        curl_easy_setopt(curl, CURLOPT_MIMEPOST, form);
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &responseString);
-
-        // Perform the request
-        res = curl_easy_perform(curl);
-        if (res != CURLE_OK) {
-            cerr << "CURL error (transcription): " << curl_easy_strerror(res) << endl;
-        }
-
-        // Clean up
-        curl_mime_free(form);
-        curl_slist_free_all(headers);
-        curl_easy_cleanup(curl);
-    }
-    curl_global_cleanup();
-    return responseString;
-}
-
-/**
- * Function to escape JSON special characters in a string
- * 
- * This function escapes special characters in a string to make it safe for inclusion in JSON.
- * It handles characters like quotes, backslashes, and control characters.
- * 
- * @param input The string to escape
- * @return The escaped string
- */
-string escapeJsonString(const string &input) {
-    ostringstream escaped;
-    for (char c : input) {
-        switch (c) {
-            case '"': escaped << "\\\""; break;
-            case '\\': escaped << "\\\\"; break;
-            case '\b': escaped << "\\b"; break;
-            case '\f': escaped << "\\f"; break;
-            case '\n': escaped << "\\n"; break;
-            case '\r': escaped << "\\r"; break;
-            case '\t': escaped << "\\t"; break;
-            default:
-                if ('\x00' <= c && c <= '\x1f') {
-                    escaped << "\\u" << hex << setw(4) << setfill('0') << (int)c;
-                } else {
-                    escaped << c;
-                }
-        }
-    }
-    return escaped.str();
-}
-
-/**
- * Function to communicate with the OpenAI Chat Completions API for categorizing transcription
- * 
- * This function sends the transcription text to the OpenAI GPT-4o model for analysis.
- * It requests categorization into various sections based on the Notion Voice Notes structure.
- * 
- * @param transcription The transcription text to analyze
- * @param apiKey OpenAI API key for authentication
- * @return The API response containing the categorized content
- */
-string categorizeWithOpenAI(const string& transcription, const string& apiKey) {
-    CURL* curl;
-    CURLcode res;
-    string responseString;
-    curl_global_init(CURL_GLOBAL_DEFAULT);
-    curl = curl_easy_init();
-    if (curl) {
-        // Escape transcription text for JSON safety
-        string escapedTranscription = escapeJsonString(transcription);
-
-        // Define the summary options from the Notion Voice Notes configuration
-        vector<string> summaryOptions = {
-            "Summary",
-            "Main Points",
-            "Action Items",
-            "References",
-            "Follow-up Questions",
-            "Stories",
-            "Arguments",
-            "Sentiment"
-        };
-
-        // Build the summary options string
-        string summaryOptionsStr = "";
-        for (size_t i = 0; i < summaryOptions.size(); ++i) {
-            summaryOptionsStr += summaryOptions[i];
-            if (i < summaryOptions.size() - 1) {
-                summaryOptionsStr += ", ";
-            }
-        }
-
-        // Build the JSON payload using a raw string literal for clarity
-        string data = R"({
-            "model": "gpt-4o",
-            "messages": [
-                {
-                    "role": "system",
-                    "content": "You are an assistant that analyzes voice recordings and outputs categorized sections in JSON format for Notion database integration."
-                },
-                {
-                    "role": "user",
-                    "content": "Analyze the following transcription and categorize it into these sections: )" + summaryOptionsStr + R"(. Generate an AI title for the note. For Type, suggest a category like 'AI Transcription', 'Meeting Notes', etc. For Duration, provide a time format like '00:07:26'. Calculate the Duration (Seconds) as a number. Include an AI Cost estimate (a small dollar amount). Also include an Icon field with the value '🤖'. Format all lists as arrays. Provide the output in clean JSON format with no markdown formatting.\n\nTranscription: )" + escapedTranscription + R"("
-                }
-            ]
-        })";
-
-        // Set up the headers
-        struct curl_slist* headers = nullptr;
-        headers = curl_slist_append(headers, ("Authorization: Bearer " + apiKey).c_str());
-        headers = curl_slist_append(headers, "Content-Type: application/json");
-
-        // Set the API endpoint for chat completions
-        curl_easy_setopt(curl, CURLOPT_URL, "https://api.openai.com/v1/chat/completions");
-        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-        curl_easy_setopt(curl, CURLOPT_POST, 1L);
-        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, data.c_str());
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &responseString);
-
-        // Perform the request
-        res = curl_easy_perform(curl);
-        if (res != CURLE_OK) {
-            cerr << "CURL error (chat completions): " << curl_easy_strerror(res) << endl;
-        }
-
-        // Clean up
-        curl_slist_free_all(headers);
-        curl_easy_cleanup(curl);
-    }
-    curl_global_cleanup();
-    return responseString;
-}
-
-/**
- * Function to ensure the Notion database has the required properties
- * 
- * This function:
- * 1. Retrieves the current structure of the Notion database
- * 2. Identifies the title property
- * 3. Checks for missing properties or properties with incorrect types
- * 4. Adds any missing properties with the correct types
- * 
- * Required properties include:
- * - Title property (for Summary)
- * - Type (select)
- * - Duration (rich_text)
- * - At Cost (number)
- * - Duration (Seconds) (number)
- * - Date (date)
- * - Main Points, Action Items, etc. (rich_text)
- * 
- * @param notionDatabaseId ID of the Notion database
- * @param notionApiKey Notion API key for authentication
- * @return true if the database has all required properties, false otherwise
- */
-bool ensureNotionDatabaseProperties(const string &notionDatabaseId,const string &notionApiKey) {
-    // Get the current database structure
-    CURL *curl;
-    CURLcode res;
-    string responseString;
-    
-    curl_global_init(CURL_GLOBAL_DEFAULT);
-    curl = curl_easy_init();
-    if (!curl) {
-        cerr << "Failed to initialize CURL for database retrieval" << endl;
-        return false;
-    }
-    
-    // Set up the required headers
-    struct curl_slist *headers = nullptr;
-    headers = curl_slist_append(headers, ("Authorization: Bearer " + notionApiKey).c_str());
-    headers = curl_slist_append(headers, "Content-Type: application/json");
-    headers = curl_slist_append(headers, "Notion-Version: 2022-06-28");
-    
-    // Set the Notion API endpoint for retrieving the database
-    string url = "https://api.notion.com/v1/databases/" + notionDatabaseId;
-    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-    curl_easy_setopt(curl, CURLOPT_WRITEDATA, &responseString);
-    
-    // Perform the request
-    res = curl_easy_perform(curl);
-    if (res != CURLE_OK) {
-        cerr << "CURL error (database retrieval): " << curl_easy_strerror(res) << endl;
-        curl_slist_free_all(headers);
-        curl_easy_cleanup(curl);
-        curl_global_cleanup();
-        return false;
-    }
-    
-    // Parse the response to check existing properties
-    try {
-        auto dbJson = json::parse(responseString);
-        
-        // Check if there was an error
-        if (dbJson.contains("object") && dbJson["object"] == "error") {
-            cerr << "Notion API error (database retrieval): " << dbJson["message"].get<string>() << endl;
-            cout << "Notion API response:" << endl << responseString << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return false;
-        }
-        
-        // Get existing properties
-        map<string, string> existingProps;
-        if (dbJson.contains("properties") && dbJson["properties"].is_object()) {
-            for (auto& [key, value] : dbJson["properties"].items()) {
-                if (value.contains("type")) {
-                    existingProps[key] = value["type"];
-                }
-            }
-        }
-        
-        // Find the existing title property
-        string titlePropName = "";
-        for (auto& [key, value] : existingProps) {
-            if (value == "title") {
-                titlePropName = key;
-                break;
-            }
-        }
-        
-        if (titlePropName.empty()) {
-            cerr << "No title property found in the database" << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return false;
-        }
-        
-        cout << "Found title property: " << titlePropName << endl;
-        
-        // Store the title property name for later use
-        g_titlePropertyName = titlePropName;
-        
-        // Define required properties with their types based on Notion Voice Notes configuration
-        map<string, string> requiredProps = {
-            {"Main Points", "rich_text"},
-            {"Action Items", "rich_text"},
-            {"Follow-up Questions", "rich_text"},
-            {"Stories", "rich_text"},
-            {"References", "rich_text"},
-            {"Arguments", "rich_text"},
-            {"Sentiment", "rich_text"},
-            {"Type", "select"},
-            {"Duration", "rich_text"},
-            {"AI Cost", "number"},
-            {"Duration (Seconds)", "number"},
-            {"Date", "date"},
-            {"Icon", "rich_text"}
-        };
-        
-        // Check which properties are missing or have the wrong type
-        vector<pair<string, string>> missingProps;
-        for (const auto& [propName, propType] : requiredProps) {
-            if (existingProps.find(propName) == existingProps.end()) {
-                missingProps.push_back({propName, propType});
-            } else if (existingProps[propName] != propType) {
-                cerr << "Warning: Property '" << propName << "' exists but has type '" 
-                     << existingProps[propName] << "' instead of '" << propType << "'" << endl;
-            }
-        }
-        
-        // If no properties are missing, return success
-        if (missingProps.empty()) {
-            cout << "All required properties exist in the database" << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return true;
-        }
-        
-        // Build the update payload
-        json updatePayload;
-        json properties;
-        
-        for (const auto& [propName, propType] : missingProps) {
-            if (propType == "rich_text") {
-                properties[propName] = {{"type", "rich_text"}, {"rich_text", json::object()}};
-            } else if (propType == "select") {
-                properties[propName] = {{"type", "select"}, {"select", json::object()}};
-            } else if (propType == "number") {
-                properties[propName] = {{"type", "number"}, {"number", json::object()}};
-            } else if (propType == "date") {
-                properties[propName] = {{"type", "date"}, {"date", json::object()}};
-            }
-        }
-        
-        updatePayload["properties"] = properties;
-        string updatePayloadStr = updatePayload.dump();
-        
-        // Reset CURL for the update request
-        curl_easy_reset(curl);
-        
-        // Set up the PATCH request
-        curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-        curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, "PATCH");
-        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, updatePayloadStr.c_str());
-        
-        // Clear the response string
-        responseString.clear();
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &responseString);
-        
-        // Perform the update request
-        res = curl_easy_perform(curl);
-        if (res != CURLE_OK) {
-            cerr << "CURL error (database update): " << curl_easy_strerror(res) << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return false;
-        }
-        
-        // Check the update response
-        try {
-            auto updateJson = json::parse(responseString);
-            if (updateJson.contains("object") && updateJson["object"] == "error") {
-                cerr << "Notion API error (database update): " << updateJson["message"].get<string>() << endl;
-                cout << "Notion API response:" << endl << responseString << endl;
-                curl_slist_free_all(headers);
-                curl_easy_cleanup(curl);
-                curl_global_cleanup();
-                return false;
-            }
-            
-            cout << "Database properties updated successfully" << endl;
-        } catch (const exception& e) {
-            cerr << "Error parsing database update response: " << e.what() << endl;
-            cout << "Raw response:" << endl << responseString << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return false;
-        }
-        
-    } catch (const exception& e) {
-        cerr << "Error parsing database response: " << e.what() << endl;
-        cout << "Raw response:" << endl << responseString << endl;
-        curl_slist_free_all(headers);
-        curl_easy_cleanup(curl);
-        curl_global_cleanup();
-        return false;
-    }
-    
-    // Clean up
-    curl_slist_free_all(headers);
-    curl_easy_cleanup(curl);
-    curl_global_cleanup();
-    
-    return true;
-}
-
-/**
- * Function to send a parsed JSON object into a Notion database
- * 
- * This function:
- * 1. Ensures the database has the required properties
- * 2. Builds a JSON payload according to Notion's API requirements
- * 3. Handles different property types (title, select, number, date, rich_text)
- * 4. Converts arrays to comma-separated strings without square brackets
- * 5. Sends the data to the Notion API
- * 
- * @param data The JSON data to send to Notion
- * @param notionDatabaseId ID of the Notion database
- * @param notionApiKey Notion API key for authentication
- * @return true if the data was successfully sent, false otherwise
- */
-bool sendToNotion(const nlohmann::json &data,const string &notionDatabaseId,const string &notionApiKey) {
-    // First, ensure the database has the required properties
-    if (!ensureNotionDatabaseProperties(notionDatabaseId, notionApiKey)) {
-        cerr << "Failed to ensure database properties" << endl;
-        return false;
-    }
-    // Build the JSON payload according to Notion's API requirements.
-    // The payload includes:
-    //   - A "parent" key specifying the database_id.
-    //   - A "properties" key that maps each key from your parsed JSON
-    //     into a Notion property.
-    nlohmann::json payload;
-    payload["parent"] = {{"database_id", notionDatabaseId}};
-    
-    nlohmann::json properties;
-    // Iterate over the key/value pairs in the input JSON.
-    for (auto& [key, value] : data.items()) {
-        // Handle each property based on its expected type in Notion
-        if (key == "AI_Title" || key == "Title") {
-            // Use the title property name from the database
-            string content = value.is_string() ? value.get<string>() : value.dump();
-            properties[g_titlePropertyName] = {
-                {"title", nlohmann::json::array({
-                    {{"text", {{"content", content}}}}
-                })}
-            };
-        } else if (key == "Summary") {
-            // Use the title property name from the database if AI_Title is not present
-            if (!data.contains("AI_Title") && !data.contains("Title")) {
-                string content = value.is_string() ? value.get<string>() : value.dump();
-                properties[g_titlePropertyName] = {
-                    {"title", nlohmann::json::array({
-                        {{"text", {{"content", content}}}}
-                    })}
-                };
-            }
-        } else if (key == "Type") {
-            // Type is a select property
-            string content = value.is_string() ? value.get<string>() : value.dump();
-            properties[key] = {
-                {"select", {{"name", content}}}
-            };
-        } else if (key == "At Cost" || key == "AI Cost") {
-            // AI Cost is a number property
-            // If the value is null or not a number, set it to null
-            string propName = "AI Cost"; // Use the Notion Voice Notes property name
-            if (value.is_null()) {
-                properties[propName] = {{"number", nullptr}};
-            } else if (value.is_number()) {
-                properties[propName] = {{"number", value}};
-            } else {
-                // Try to convert string to number
-                try {
-                    double numValue = stod(value.is_string() ? value.get<string>() : value.dump());
-                    properties[propName] = {{"number", numValue}};
-                } catch (...) {
-                    properties[propName] = {{"number", nullptr}};
-                }
-            }
-        } else if (key == "Duration (Seconds)") {
-            // Duration (Seconds) is a number property
-            if (value.is_number()) {
-                properties[key] = {{"number", value}};
-            } else {
-                // Try to convert string to number
-                try {
-                    double numValue = stod(value.is_string() ? value.get<string>() : value.dump());
-                    properties[key] = {{"number", numValue}};
-                } catch (...) {
-                    properties[key] = {{"number", 0}};
-                }
-            }
-        } else if (key == "Date") {
-            // Date is a date property
-            if (value.is_null()) {
-                properties[key] = {{"date", nullptr}};
-            } else {
-                string dateStr = value.is_string() ? value.get<string>() : value.dump();
-                if (dateStr == "null" || dateStr.empty()) {
-                    properties[key] = {{"date", nullptr}};
-                } else {
-                    properties[key] = {{"date", {{"start", dateStr}}}};
-                }
-            }
-        } else {
-            // All other properties are rich_text
-            if (value.is_array()) {
-                // Convert array to string without square brackets
-                ostringstream contentStream;
-                for (size_t i = 0; i < value.size(); ++i) {
-                    string itemText = value[i].is_string() ? value[i].get<string>() : value[i].dump();
-                    contentStream << itemText;
-                    if (i < value.size() - 1) {
-                        contentStream << ", ";
-                    }
-                }
-                string content = contentStream.str();
-                properties[key] = {
-                    {"rich_text", nlohmann::json::array({
-                        {{"text", {{"content", content}}}}
-                    })}
-                };
-            } else {
-                // Handle non-array values as before
-                string content = value.is_string() ? value.get<string>() : value.dump();
-                properties[key] = {
-                    {"rich_text", nlohmann::json::array({
-                        {{"text", {{"content", content}}}}
-                    })}
-                };
-            }
-        }
-    }
-    payload["properties"] = properties;
-    
-    // Convert the payload JSON to a string
-    string payloadStr = payload.dump();
-    
-    // Set up CURL for the HTTP POST request to Notion's API.
-    CURL *curl;
-    CURLcode res;
-    string responseString;
-    
-    curl_global_init(CURL_GLOBAL_DEFAULT);
-    curl = curl_easy_init();
-    if (curl) {
-        // Set up the required headers.
-        struct curl_slist *headers = nullptr;
-        headers = curl_slist_append(headers, ("Authorization: Bearer " + notionApiKey).c_str());
-        headers = curl_slist_append(headers, "Content-Type: application/json");
-        headers = curl_slist_append(headers, "Notion-Version: 2022-06-28"); // Adjust if needed
-
-        // Set the Notion API endpoint for creating a page
-        curl_easy_setopt(curl, CURLOPT_URL, "https://api.notion.com/v1/pages");
-        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-        curl_easy_setopt(curl, CURLOPT_POST, 1L);
-        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, payloadStr.c_str());
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &responseString);
-        
-        // Perform the request
-        res = curl_easy_perform(curl);
-        if (res != CURLE_OK) {
-            cerr << "CURL error (Notion API): " << curl_easy_strerror(res) << endl;
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(curl);
-            curl_global_cleanup();
-            return false;
-        }
-        
-        // Check the response for errors
-        try {
-            auto responseJson = json::parse(responseString);
-            if (responseJson.contains("object") && responseJson["object"] == "error") {
-                cerr << "Notion API error: " << responseJson["message"].get<string>() << endl;
-                cout << "Notion API response:" << endl << responseString << endl;
-                curl_slist_free_all(headers);
-                curl_easy_cleanup(curl);
-                curl_global_cleanup();
-                return false;
-            }
-        } catch (const exception& e) {
-            // If we can't parse the response, just print it
-            cout << "Notion API response:" << endl << responseString << endl;
-        }
-        
-        // Clean up
-        curl_slist_free_all(headers);
-        curl_easy_cleanup(curl);
-    }
-    curl_global_cleanup();
-    
-    return true;
-}
 
 /**
  * Main function - Entry point of the application
