+++ analysis.h
@@ -1,0 +1,8 @@
+#ifndef ANALYSIS_H
+#define ANALYSIS_H
+
+#include <string>
+
+std::string categorizeWithOpenAI(const std::string &transcription, const std::string &apiKey);
+
+#endif // ANALYSIS_H
