+++ transcription.h
@@ -1,0 +1,8 @@
+#ifndef TRANSCRIPTION_H
+#define TRANSCRIPTION_H
+
+#include <string>
+
+std::string transcribeAudio(const std::string &filePath, const std::string &apiKey);
+
+#endif // TRANSCRIPTION_H
